Added a python file that looks for duplicated ID's for debugging
For future cases, to quickly find the duplicated IDs, A python code is added to check if there is a duplicated ID. This is integrated into the Github workflow to check every time if there is a duplicated ID or not. This helps to quickly identify the duplicated ID and remove it before the build starts.